basic lighting

+++ src/vecs/vec2.rs
@@ -1,10 +1,11 @@
-pub struct Vec2f {
+#[derive(Debug, Clone, Copy)]
+pub struct Vec2i {
     pub x: i32,
     pub y: i32,
 }
 
-impl Vec2f {
-    pub fn new(x: i32, y: i32) -> Vec2f {
-        Vec2f { x, y }
+impl Vec2i {
+    pub fn new(x: i32, y: i32) -> Vec2i {
+        Vec2i { x, y }
     }
 }
